+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed an issue that would cause unnamed files to no longer appear in the raw client files list.
 - Fixed an issue that would prevent Earthen characters from importing from Battle.net.
 - Fixed item model exports missing material textures.
+- Fixed an issue that prevents copying rows from the data table while a filter is active.
 
 ## 0.2.14 (19-03-2026)
 - Added "Creatures" tab, allowing export of NPCs with both custom and character models.